extracted segment changer + fixed use of property
fixes #28
