+++ german_freeval/input/property_builder.py
@@ -1,4 +1,4 @@
-import numpy as np
+from typing import Any
 
 from german_freeval.macro.property import Property
 
@@ -8,20 +8,20 @@ class PropertyBuilder:
     type: str
     values: dict
 
-    def __init__(self, name: str, type: str, startvalue) -> None:
+    def __init__(self, name: str, type: str, initial_values: Any | list | dict) -> None:
         self.name = name
         self.type = type
-        self.values = {0: startvalue}
+
+        match initial_values:
+            case list():
+                self.values = {k: v for k, v in enumerate(initial_values)}
+            case dict():
+                self.values = initial_values.copy()
+            case _:
+                self.values = {0: initial_values}
 
     def add_period_value(self, period: int, value):
         self.values[period] = value
-
-    def get_value(self, period: int):
-        periods = np.sort(self.values.keys())
-        for i in range(len(periods)):
-            if periods[i] - period > 0:
-                key_period = periods[i - 1]
-        return self.values[key_period]
 
     def build(self, n_periods: int):
         values = [0] * n_periods
